refactor(frontend): superadmin ページの Hook+View 違反を解消 (#246)
- features/manage-data-migration/ を新規作成
  - useDataMigrationPage hook（移行状態・組織一覧・割当処理）
  - DataMigrationView コンポーネント（純粋 props 受け取り）
- features/manage-organizations/ を新規作成
  - useManageOrganizationsPage hook（一覧・作成・削除・UI 状態）
  - useManageOrganizationDetailPage hook（詳細・更新・削除・Export/Import）
  - ManageOrganizationsView コンポーネント（CreateOrganizationForm は内部非公開コンポーネント）
  - ManageOrganizationDetailView コンポーネント（useRef のみ保持）
- 各ページを hook 呼び出し + view spread の 3 行に削減
