+++ frontend/src/features/manage-data-migration/index.ts
@@ -1,0 +1,3 @@
+export { DataMigrationView } from './ui/DataMigrationView'
+export { useDataMigrationPage } from './hooks/useDataMigrationPage'
+export type { DataMigrationPageState } from './hooks/useDataMigrationPage'
